move tests around to show new workflow
